pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-ai-toolkit"
version = "0.1.0"
description = "Unified toolkit for video AI: segmentation, depth, matting, and inpainting"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "AFTRS Studio"}
]
keywords = ["video", "ai", "segmentation", "depth", "matting", "inpainting", "sam2", "propainter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "click>=8.1.0",
    "opencv-python>=4.8.0",
    "ffmpeg-python>=0.2.0",
    "huggingface-hub>=0.20.0",
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
segment = [
    "segment-anything-2",
    "groundingdino",
]
depth = [
    "depth-anything-v2",
]
matting = [
    "robust-video-matting",
]
inpaint = [
    "propainter",
]
all = [
    "video-ai-toolkit[segment,depth,matting,inpaint]",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
vidtool = "video_toolkit.cli:cli"

[project.urls]
Homepage = "https://github.com/aftrs-studio/video-ai-toolkit"
Repository = "https://github.com/aftrs-studio/video-ai-toolkit.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["video_toolkit*"]

[tool.black]
line-length = 88
target-version = ["py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
